arm/aarch64: vp9: Fix vertical alignment
Align the second/third operands as they usually are.

Due to the wildly varying sizes of the written out operands
in aarch64 assembly, the column alignment is usually not as clear
as in arm assembly.

This is cherrypicked from libav commit
7995ebfad12002033c73feed422a1cfc62081e8f.

Signed-off-by: Martin Storsjö <martin@martin.st>
(cherry picked from commit 21c89f3a26bb1331381b90e653277585447cfbb3)
Signed-off-by: Michael Niedermayer <michael@niedermayer.cc>